% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcxn.R
\docType{class}
\name{pcxn-class}
\alias{pcxn-class}
\title{A pcxn object produced by pcxn_explore() or pcxn_analyze(). It holds the 
corresponding analysis, the data produced by the analysis and the geneset 
groups involved.}
\value{
pcxn object with a type, data and geneset_groups field
}
\description{
A pcxn object produced by pcxn_explore() or pcxn_analyze(). It holds the 
corresponding analysis, the data produced by the analysis and the geneset 
groups involved.
}
\section{Slots}{

\describe{
\item{\code{type}}{character.}

\item{\code{data}}{matrix.}

\item{\code{geneset_groups}}{list.}
}}

\examples{
# Create and show a pcxn object
pcxn <- pcxn_explore("pathprint","Alzheimer's disease (KEGG)", 10, 
0.05, 0.05)

pcxn
}
