\name{pcxn}
\alias{Introduction to pcxn}
\alias{pcxn}
\docType{package}
\title{Pathway correlation coefficients and p values between pre-defined gene
sets}
\description{
PCxN database contains correlation coefficients and p values between pre-defined
gene sets within MSigDB H hallmark gene sets, MSigDB C2 CP (Canonical pathways),
MSigDB C5 GO BP gene sets and Pathprint respectively
}
\details{
\tabular{ll}{
Package: \tab pcxn\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2017-08-20\cr
License: \tab GPL\cr
}
}
\author{
Sokratis Kariotis
\cr Maintainer: Sokratis Kariotis \email{s.kariotis@sheffield.ac.uk}
}
\references{
Pita-Juarez Y.,Altschuler G.,Kariotis S.,Wei W.,Koler K.,Tanzi R.
and W. A. Hide (2017). "The Pathway Coexpression Network: Revealing Pathway
Relationships."
}
\keyword{ package }

\examples{
library(pcxnData)

# load  the data
ds = c("cp_gs_v5.1", "gobp_gs_v5.1", "h_gs_v5.1","pathprint.Hs.gs",
    "pathCor_CPv5.1_dframe","pathCor_GOBPv5.1_dframe","pathCor_Hv5.1_dframe",
    "pathCor_pathprint_v1.2.3_dframe")
data(list = ds)

# Explore the static extendable network by focusing on single pathways and their
# 10 most correlated neighbours in the pathprint collection
pcxn.obj <- pcxn.explore(collection = "pathprint",
                    query_geneset = "Alzheimer's disease (KEGG)",
                    top = 10,
                    min_abs_corr = 0.05,
                    max_pval = 0.05)

# Analyse relationships between groups of pathways shown to be enriched in the
# collection by gene set enrichment
pcxn.obj <- pcxn.analyze(collection = "pathprint",
            phenotype_0_genesets = c("ABC transporters (KEGG)",
                                    "ACE Inhibitor Pathway (Wikipathways)",
                                    "AR down reg. targets (Netpath)"),
            phenotype_1_genesets = c("DNA Repair (Reactome)"),
            top = 10,
            min_abs_corr = 0.05,
            max_pval = 0.05 )

# Generate the heatmap for any pcxn object generated by the
# explore or analyze function
hm <- pcxn.heatmap(pcxn.obj , cluster_method = "complete")

# Get the gene members (Entrez Ids and names) of any pathway/geneset in the
# data
genesets_list <- pcxn.gene_members(pathway_name = "Alzheimer's disease (KEGG)")

# Create a network for any pcxn object generated by the
# explore or analyze function
# network <- pcxn.network(pcxn.obj)

}
