% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcxn.obj.R
\docType{class}
\name{pcxn.obj-class}
\alias{pcxn.obj-class}
\title{PCxN object which holds the type of analysis it came from, the data
produced by the analysis and the geneset groups involved. It can be produced 
by the pcxn.explore() or pcxn.analyze() functions.}
\value{
pcxn.obj with a type, data and geneset_grous field
}
\description{
PCxN object which holds the type of analysis it came from, the data produced
by the analysis and the geneset groups involved. It can be produced by the
pcxn.explore() or pcxn.analyze() functions.
}
\section{Slots}{

\describe{
\item{\code{type}}{character.}

\item{\code{data}}{matrix.}

\item{\code{geneset_groups}}{list.}
}}

\examples{
\dontrun{
pcxn.obj
}

}
