% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcxn_network.R
\name{pcxn_network}
\alias{pcxn_network}
\title{Create a network of a pcxn object}
\usage{
pcxn_network(object)
}
\arguments{
\item{object}{pcxn object created by explore or analyze functions}
}
\value{
draws a tkplot object and saves a graph object 
representing the network
}
\description{
Create a network of a pcxn object
}
\examples{
# Create a network of a pcxn object
object <- pcxn_explore("pathprint","Alzheimer's disease (KEGG)", 
10, 0.05, 0.05)

# network <- pcxn_network(object)

}
