% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcxn.network.R
\name{pcxn.network}
\alias{pcxn.network}
\title{Create a network of a pcxn object}
\usage{
pcxn.network(df)
}
\arguments{
\item{df}{pcxn object created by explore or analyze functions}
}
\value{
tkplot object representing the network
}
\description{
Create a network of a pcxn object
}
\examples{
# Create a network of a pcxn object
df <- pcxn.explore("pathprint","Alzheimer's disease (KEGG)", 10, 0.05, 0.05)
# network <- pcxn.network(df)
}
