% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcxn_network.R
\name{pcxn_network}
\alias{pcxn_network}
\title{Create a network of a pcxn object}
\usage{
pcxn_network(df)
}
\arguments{
\item{df}{pcxn object created by explore or analyze functions}
}
\value{
tkplot object representing the network
}
\description{
Create a network of a pcxn object
}
\examples{
# Create a network of a pcxn object
df <- pcxn_explore("pathprint","Alzheimer's disease (KEGG)", 10, 0.05, 0.05)

# network <- pcxn_network(df)
}
